Add basic files
README, gitignore and license files were not added on the inital commit (2d55ffe88fecae5588634e2ce215579df97a2093) because files for the project were already present locally but a repository had not been initalized. The creation of these files on the initial commit may have caused to avoid errors.

Previous test with standard.js was removed and a start command was added for convinience.